pyproject.toml
[project]
name = "agenticai-lang-graph"
version = "0.1.0"
description = "Agentic AI Lang Graph"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "easyocr>=1.7.2",
    "langgraph>=1.0.3",
    "librosa>=0.11.0",
    "moviepy>=2.2.1",
    "numpy>=2.3.5",
    "openai-whisper>=20250625",
    "opencv-contrib-python>=4.11.0.86",
    "scipy>=1.16.3",
    "yt-dlp>=2025.11.12",
    "face-alignment>=1.4.1",
    "duckduckgo-search>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "sentence-transformers>=2.2.0",
    "spacy>=3.7.0",
    "redis>=5.0.0",
    "python-whois>=0.8.0",
    "aiohttp>=3.9.0",
    "prometheus-client>=0.19.0",
    "sixdrepnet>=0.1.6",
    "pip>=24.0",
    "spacy-transformers>=1.3.4",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "graphviz>=0.21",
]
